move netcdf creation to separate file
